--- src/main/java/ru/job4j/dream/servlet/candidate/CandidateBrowseServlet.java
@@ -1,17 +1,0 @@
-package ru.job4j.dream.servlet.candidate;
-
-import ru.job4j.dream.store.Store;
-
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-
-public class CandidateBrowseServlet extends HttpServlet {
-    @Override
-    protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        req.setAttribute("candidates", Store.instOf().findAllCandidates());
-        req.getRequestDispatcher("candidate/candidates.jsp").forward(req, resp);
-    }
-}
