+++ src/main/java/ru/job4j/dream/servlet/candidate/CandidateServlet.java
@@ -10,7 +10,7 @@ import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.util.Objects;
 
-public class CandidateEditServlet extends HttpServlet {
+public class CandidateServlet extends HttpServlet {
     @Override
     protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
         req.setCharacterEncoding("UTF-8");
@@ -27,5 +27,11 @@ public class CandidateEditServlet extends HttpServlet {
         }
 
         resp.sendRedirect(req.getContextPath() + "/candidates.do");
+    }
+
+    @Override
+    protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+        req.setAttribute("candidates", Store.instOf().findAllCandidates());
+        req.getRequestDispatcher("candidate/candidates.jsp").forward(req, resp);
     }
 }
