--- src/main/java/ru/job4j/dream/servlet/post/PostBrowseServlet.java
@@ -1,17 +1,0 @@
-package ru.job4j.dream.servlet.post;
-
-import ru.job4j.dream.store.Store;
-
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-
-public class PostBrowseServlet extends HttpServlet {
-    @Override
-    protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        req.setAttribute("posts", Store.instOf().findAllPosts());
-        req.getRequestDispatcher("post/posts.jsp").forward(req, resp);
-    }
-}
