+++ src/main/java/ru/job4j/dream/servlet/post/PostServlet.java
@@ -10,9 +10,7 @@ import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.util.Objects;
 
-public class PostEditServlet extends HttpServlet {
-
-
+public class PostServlet extends HttpServlet {
     @Override
     protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
         req.setCharacterEncoding("UTF-8");
@@ -27,5 +25,11 @@ public class PostEditServlet extends HttpServlet {
         }
 
         resp.sendRedirect(req.getContextPath() + "/posts.do");
+    }
+
+    @Override
+    protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+        req.setAttribute("posts", Store.instOf().findAllPosts());
+        req.getRequestDispatcher("post/posts.jsp").forward(req, resp);
     }
 }
