+++ src/main/webapp/WEB-INF/web.xml
@@ -1,49 +1,31 @@
 <!DOCTYPE web-app PUBLIC
- "-//Sun Microsystems, Inc.//DTD Web Application 2.3//EN"
- "http://java.sun.com/dtd/web-app_2_3.dtd" >
+        "-//Sun Microsystems, Inc.//DTD Web Application 2.3//EN"
+        "http://java.sun.com/dtd/web-app_2_3.dtd" >
 
 <web-app>
-  <display-name>Archetype Created Web Application</display-name>
-  <servlet>
-    <servlet-name>IndexServlet</servlet-name>
-    <servlet-class>ru.job4j.dream.servlet.index.IndexServlet</servlet-class>
-  </servlet>
-  <servlet>
-    <servlet-name>PostEditServlet</servlet-name>
-    <servlet-class>ru.job4j.dream.servlet.post.PostEditServlet</servlet-class>
-  </servlet>
-  <servlet>
-    <servlet-name>CandidateEditServlet</servlet-name>
-    <servlet-class>ru.job4j.dream.servlet.candidate.CandidateEditServlet</servlet-class>
-  </servlet>
-  <servlet>
-    <servlet-name>PostBrowseServlet</servlet-name>
-    <servlet-class>ru.job4j.dream.servlet.post.PostBrowseServlet</servlet-class>
-  </servlet>
-  <servlet>
-    <servlet-name>CandidateBrowseServlet</servlet-name>
-    <servlet-class>ru.job4j.dream.servlet.candidate.CandidateBrowseServlet</servlet-class>
-  </servlet>
-  <servlet-mapping>
-    <servlet-name>IndexServlet</servlet-name>
-    <url-pattern>/index.do</url-pattern>
-  </servlet-mapping>
-  <servlet-mapping>
-    <servlet-name>PostBrowseServlet</servlet-name>
-    <url-pattern>/posts.do</url-pattern>
-  </servlet-mapping>
-  <servlet-mapping>
-    <servlet-name>PostEditServlet</servlet-name>
-    <url-pattern>/post/save</url-pattern>
-  </servlet-mapping>
-  <servlet-mapping>
-    <servlet-name>CandidateEditServlet</servlet-name>
-    <url-pattern>/candidate/save</url-pattern>
-  </servlet-mapping>
-  <servlet-mapping>
-    <servlet-name>CandidateBrowseServlet</servlet-name>
-    <url-pattern>/candidates.do</url-pattern>
-  </servlet-mapping>
-
-
+    <display-name>Archetype Created Web Application</display-name>
+    <servlet>
+        <servlet-name>IndexServlet</servlet-name>
+        <servlet-class>ru.job4j.dream.servlet.index.IndexServlet</servlet-class>
+    </servlet>
+    <servlet>
+        <servlet-name>PostBrowseServlet</servlet-name>
+        <servlet-class>ru.job4j.dream.servlet.post.PostServlet</servlet-class>
+    </servlet>
+    <servlet>
+        <servlet-name>CandidateBrowseServlet</servlet-name>
+        <servlet-class>ru.job4j.dream.servlet.candidate.CandidateServlet</servlet-class>
+    </servlet>
+    <servlet-mapping>
+        <servlet-name>IndexServlet</servlet-name>
+        <url-pattern>/index.do</url-pattern>
+    </servlet-mapping>
+    <servlet-mapping>
+        <servlet-name>PostBrowseServlet</servlet-name>
+        <url-pattern>/posts.do</url-pattern>
+    </servlet-mapping>
+    <servlet-mapping>
+        <servlet-name>CandidateBrowseServlet</servlet-name>
+        <url-pattern>/candidates.do</url-pattern>
+    </servlet-mapping>
 </web-app>
