+++ src/main/webapp/post/edit.jsp
@@ -28,7 +28,7 @@
                 Новая вакансия.
             </div>
             <div class="card-body">
-                <form action="<%=request.getContextPath()%>/post/save?id=<%=request.getParameter("id")%>" method="post">
+                <form action="<%=request.getContextPath()%>/posts.do?id=<%=request.getParameter("id")%>" method="post">
                     <div class="form-group">
                         <label>Имя</label>
                         <input type="text" class="form-control" name="name" value="<%=request.getParameter("name")%>">
